Fix (MSVC) compiler warnings reg. `size_t` vs. `int`
as per code review, https://github.com/musescore/MuseScore/pull/11197#pullrequestreview-943197231

+++ src/engraving/libmscore/note.h
@@ -253,7 +253,7 @@ public:
     // Score Tree functions
     EngravingObject* scanParent() const override;
     EngravingObject* scanChild(int idx) const override;
-    int scanChildCount() const override;
+    size_t scanChildCount() const override;
 
     void undoUnlink() override;
