+++ src/engraving/libmscore/system.h
@@ -139,7 +139,7 @@ public:
     // Score Tree functions
     EngravingObject* scanParent() const override;
     EngravingObject* scanChild(int idx) const override;
-    int scanChildCount() const override;
+    size_t scanChildCount() const override;
 
     System* clone() const override { return new System(*this); }
