+++ src/engraving/libmscore/textframe.h
@@ -47,7 +47,7 @@ public:
     // Score Tree functions
     EngravingObject* scanParent() const override;
     EngravingObject* scanChild(int idx) const override;
-    int scanChildCount() const override;
+    size_t scanChildCount() const override;
 
     TBox* clone() const override { return new TBox(*this); }
